Refactor: Split AdminDashboard.tsx into components
Splits the AdminDashboard.tsx file into smaller, more manageable components to improve code organization and readability.
